Fixes #2. Renamed top level package to ui
1. Renamed top level package
2. Moved files in test and androidTest sourceSets to ui package
