+++ app/src/main/java/ui/MainActivity.kt
@@ -1,12 +1,13 @@
-package com.merryapps.namingtips
+package ui
 
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import com.merryapps.namingtips.R.layout
 
 class MainActivity : AppCompatActivity() {
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    setContentView(R.layout.activity_main)
+    setContentView(layout.activity_main)
   }
 }
