+++ app/src/test/java/ui/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package com.merryapps.namingtips
+package ui
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
